Added per-session lock to kernel cleanup
